refactor: extract createForumPost and splitMessage from submitApplication
createForumPost is now a standalone exported function in
src/functions/applications/createForumPost.ts so it can be reused by the
upcoming /testdata seed_application discord:true path. Behavior is unchanged.
